+++ app/src/ui/app.tsx
@@ -70,7 +70,7 @@ import { sendReady } from './main-process-proxy'
 import { TermsAndConditions } from './terms-and-conditions'
 import { ZoomInfo } from './window/zoom-info'
 import { FullScreenInfo } from './window/full-screen-info'
-import { PushBranchCommits } from './branches/PushBranchCommits'
+import { PushBranchCommits } from './branches/push-branch-commits'
 import { Branch } from '../models/branch'
 import { CLIInstalled } from './cli-installed'
 import { GenericGitAuthentication } from './generic-git-auth'
